separate generatePdf and generateDocx
